ocr/requirements.txt
pyTesseract
Pillow
numpy
opencv-python